Clap changed to configure_me in wired
This implements basic changes to use `configure_me` instead of `clap`
for managing the configuration. This instantly allows us to pull the
configuration from files, directories containing files and to generate
man pages or other integrations that will be developed in `configure_me`
in the future.

So far, this is missing `connect` because it wasn't used and
`configure_me` doesn't support parsing multiple arguments into arrays
yet.

Closes #9